requirements.txt
flashtext==2.7
newspaper3k==0.2.8
streamlit==1.43.0
yake==0.4.8
lxml_html_clean==0.4.1
PyPDF2==3.0.1
pdfplumber==0.11.6